Separate glue code into more modules
