+++ myst_nb/__init__.py
@@ -18,7 +18,8 @@ from .parser import (
     CellOutputBundleNode,
 )
 from .transform import CellOutputsToNodes
-from .glue.sphinx import Paste, paste_role, PasteNodesToDocutils
+from .glue.domain import Paste, paste_role
+from .glue.transform import PasteNodesToDocutils
 
 
 def static_path(app):
